Add add_mako_renderer to Configurator
This will allow an end user to register more file extensions with a
different setting "namespace". For example:

    config.add_mako_renderer('.email', settings_prefix='mail_mako.')

This will read from the config.registry.settings dict and look for
anything prefixed with mail_mako., this will then be used to set up
Mako. Now it is as simple as using:

    from pyramid.renderers import render
    result = render('password_reset.email', {username: 'John',
                                             reset_link:
                                             'http://example.net/login'
                                             })

to render HTML email's that can then be sent out.