Rename Test renderer factory to Test initialize renderer
Since we are no longer lazy loading Mako's lookup mechanism we can't
expect that calling renderer_factory will everything for us.

We have to explicitly initialize the renderer with the settings we want
by adding the settings to the current config object. The settings passed
in with info to the renderer_factory is basically ignored now, and no
longer has any effect.